refactor(core/modal): Extract WizardStepLabel to its own file
