+++ app/scripts/modules/core/src/modal/index.ts
@@ -1,6 +1,5 @@
 export * from './buttons/ModalClose';
 export * from './buttons/SubmitButton';
-export * from './wizard/WizardModal';
-export * from './wizard/WizardPage';
+export * from './wizard';
 export { V2_MODAL_WIZARD_COMPONENT, V2ModalWizard } from './wizard/v2modalWizard.component';
 export { IWizardPage, ModalWizard, IWizardPageState } from './wizard/ModalWizard';
