+++ app/scripts/modules/core/src/modal/wizard/index.ts
@@ -1,0 +1,3 @@
+export * from './WizardModal';
+export * from './WizardPage';
+export * from './WizardStepLabel';
